Use HTTPS for git clone and add MD5 hashing
This commit switches from git clone using ssh to git clone using https.

With SSH, the box must have ssh pub/priv keys for github private repos.
With HTTPS, we can reuse the GITHUB_ACCESS_TOKEN used for API calls.

It also changes the structure of the final file's key on S3 so that
the `/` characters cause folder-like behavior in S3 clients. And we
append a MD5 hash of the final tar.gz for validation.